+++ .env
@@ -1,0 +1,1 @@
+DATABASE_URL = postgres://hi:a@localhost:5432/sdc
