+++ package.json
@@ -21,8 +21,11 @@
   },
   "dependencies": {
     "body-parser": "^1.18.3",
+    "dotenv": "^6.1.0",
     "express": "^4.16.4",
+    "faker": "^4.1.0",
     "mongoose": "^5.3.6",
+    "pg": "^7.6.1",
     "prop-types": "^15.6.2",
     "react": "^16.6.0",
     "react-dom": "^16.6.0",
@@ -50,6 +53,7 @@
     "eslint-plugin-react": "^7.11.1",
     "html-webpack-plugin": "^3.2.0",
     "jest": "^23.6.0",
+    "make-runnable": "^1.3.6",
     "merge": "^1.2.1",
     "prettier": "1.14.3",
     "regenerator-runtime": "^0.12.1",
